requirements.txt
python-dotenv>=1.0.1
pydantic>=2.7.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.2.0
tqdm>=4.66.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pycocotools>=2.0.7
fastapi>=0.112.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pytest>=8.0.0
